Fix link when PIE is enabled by default
The code produced by the assembler isn't position-independent, which the
linker needs to be aware of. This is done with the -no-pie gcc option,
but this option is recent and isn't supported by older compilers where
PIE isn't enabled by default (and thus not needed).

As a result, add a Makefile macro to test for the support of compiler
options, and if -no-pie is supported, add it to the linker flags.